requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# Google Sheets Integration
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# AI and ML
google-generativeai==0.3.2
openai==1.3.7

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.1

# HTTP Requests and APIs
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Scheduling and Background Tasks
APScheduler==3.10.4

# Logging and Monitoring
coloredlogs==15.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
